+++ trace/driver.go
@@ -260,7 +260,9 @@ type (
 		// Safe replacement of context are provided only inside callback function
 		Context *context.Context
 	}
-	ClusterInitDoneInfo   struct{}
+	ClusterInitDoneInfo struct {
+		Error error
+	}
 	ClusterCloseStartInfo struct {
 		// Context make available context in trace callback function.
 		// Pointer to context provide replacement of context in trace callback function.