+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.13.0
+* Refactored `Connection` interface
+* Removed `CustomOption` and taking client with custom options
+* Removed `proxy` package
+* Improved `db.With()` helper for child connections creation
+* Set shared `conn.Pool` for all children `ydb.Connection`
+* Fixed bug with `RoundRobin` and `RandomChoice` balancers `Create()`
+
 ## 3.12.1
 * Added `trace.Driver.OnConnPark` event
 * Added `trace.Driver.OnConnClose` event